+++ examples/resources/coralogix_webhook/resource.tf
@@ -79,10 +79,8 @@ resource "coralogix_webhook" "demisto_webhook" {
 }
 
 resource "coralogix_webhook" "sendlog_webhook" {
-  name = "sendlog-webhook"
-  sendlog = {
-    url = "https://example-url.com/"
-  }
+  name    = "sendlog-webhook"
+  sendlog = {}
 }
 
 resource "coralogix_webhook" "event_bridge_webhook" {